feat: all update in session, course, attendee sync to Redis
