+++ src/main/java/com/waruru/areyouhere/active/domain/entity/CurrentSessionAttendanceInfo.java
@@ -7,11 +7,13 @@ import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.redis.core.RedisHash;
 
@@ -26,20 +28,26 @@ public class CurrentSessionAttendanceInfo {
     @NotNull
     private long sessionId;
 
+    @NotNull
+    private long courseId;
+
     private List<AttendeeRedisData> attendees;
 
     @NotNull
+    @Setter
     private String courseName;
 
     @NotNull
+    @Setter
     private String sessionName;
 
     private LocalDateTime createdAt;
 
     @Builder
-    public CurrentSessionAttendanceInfo(String authCode, long sessionId, List<Attendee> attendees, String courseName, String sessionName) {
+    public CurrentSessionAttendanceInfo(String authCode, long sessionId, long courseId, List<Attendee> attendees, String courseName, String sessionName) {
         this.authCode = authCode;
         this.sessionId = sessionId;
+        this.courseId = courseId;
         this.attendees = Optional.ofNullable(attendees)
                 .orElse(Collections.emptyList())
                 .stream()
@@ -48,9 +56,22 @@ public class CurrentSessionAttendanceInfo {
                         .name(attendee.getName())
                         .note(attendee.getNote())
                         .build())
-                .collect(Collectors.toList());
+                .toList();
         this.courseName = courseName;
         this.sessionName = sessionName;
         this.createdAt = LocalDateTime.now();
     }
+
+    public void updateAttendees(List<Attendee> attendees){
+        this.attendees = Optional.ofNullable(attendees)
+                .orElse(Collections.emptyList())
+                .stream()
+                .map(attendee -> AttendeeRedisData.builder()
+                        .id(attendee.getId())
+                        .name(attendee.getName())
+                        .note(attendee.getNote())
+                        .build())
+                .toList();
+    }
+
 }
