+++ src/main/java/com/waruru/areyouhere/attendance/domain/repository/AttendanceRepository.java
@@ -11,7 +11,7 @@ public interface AttendanceRepository extends JpaRepository<Attendance, Long>{
     public List<Attendance> findAttendancesBySession_Id(Long sessionId);
 
 
-    @Modifying(clearAutomatically = true)
+    @Modifying(clearAutomatically = true, flushAutomatically = true)
     @Query("delete from attendance a where a.attendee.id in :ids")
     public void deleteAllByAttendeeIds(@Param("ids") List<Long> ids);
