+++ src/main/java/com/waruru/areyouhere/course/exception/CourseActivatedSessionException.java
@@ -1,0 +1,15 @@
+package com.waruru.areyouhere.course.exception;
+
+public class CourseActivatedSessionException extends RuntimeException{
+    public CourseActivatedSessionException() {
+        super();
+    }
+
+    public CourseActivatedSessionException(String message) {
+        super(message);
+    }
+
+    public CourseActivatedSessionException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
