+++ src/main/java/com/waruru/areyouhere/course/service/CourseServiceImpl.java
@@ -1,5 +1,6 @@
 package com.waruru.areyouhere.course.service;
 
+import com.waruru.areyouhere.active.ActiveSessionService;
 import com.waruru.areyouhere.attendance.domain.repository.AttendanceRepository;
 import com.waruru.areyouhere.attendee.domain.entity.Attendee;
 import com.waruru.areyouhere.attendee.domain.repository.AttendeeBatchRepository;
@@ -10,10 +11,12 @@ import com.waruru.areyouhere.course.domain.entity.Course;
 import com.waruru.areyouhere.course.domain.repository.CourseRepository;
 import com.waruru.areyouhere.attendee.exception.AttendeesNotUniqueException;
 import com.waruru.areyouhere.course.dto.CourseData;
+import com.waruru.areyouhere.course.exception.CourseActivatedSessionException;
 import com.waruru.areyouhere.manager.domain.entity.Manager;
 import com.waruru.areyouhere.manager.domain.repository.ManagerRepository;
 import com.waruru.areyouhere.session.domain.entity.Session;
 import com.waruru.areyouhere.session.domain.repository.SessionRepository;
+import com.waruru.areyouhere.session.exception.ActivatedSessionExistsException;
 import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -36,6 +39,7 @@ public class CourseServiceImpl implements CourseService {
     private final AttendeeRepository attendeeRepository;
     private final AttendanceRepository attendanceRepository;
     private final SessionRepository sessionRepository;
+    private final ActiveSessionService activeSessionService;
 
 
     @Override
@@ -102,10 +106,15 @@ public class CourseServiceImpl implements CourseService {
 
         course.update(name, description, onlyListNameAllowed);
         courseRepository.save(course);
+        activeSessionService.updateCourseName(courseId, name);
     }
 
     @Override
     public void delete(Long managerId, Long courseId) {
+        if(activeSessionService.isSessionActivatedByCourseId(courseId)){
+            throw new ActivatedSessionExistsException("Session is activated");
+        }
+
         Course course = courseRepository.findById(courseId).
                 orElseThrow(() -> new IllegalArgumentException("Course not found"));
