+++ src/main/java/com/waruru/areyouhere/session/advice/SessionExceptionAdvice.java
@@ -1,11 +1,13 @@
 package com.waruru.areyouhere.session.advice;
 
 import static com.waruru.areyouhere.common.utils.HttpStatusResponseEntity.RESPONSE_BAD_REQUEST;
+import static com.waruru.areyouhere.common.utils.HttpStatusResponseEntity.RESPONSE_CONFLICT;
 import static com.waruru.areyouhere.common.utils.HttpStatusResponseEntity.RESPONSE_NOT_FOUND;
 import static com.waruru.areyouhere.common.utils.HttpStatusResponseEntity.RESPONSE_NO_CONTENT;
 
 import com.waruru.areyouhere.attendee.exception.AttendeeNotFoundException;
 import com.waruru.areyouhere.course.exception.CourseNotFoundException;
+import com.waruru.areyouhere.session.exception.ActivatedSessionExistsException;
 import com.waruru.areyouhere.session.exception.CurrentSessionDeactivatedException;
 import com.waruru.areyouhere.session.exception.CurrentSessionNotFoundException;
 import com.waruru.areyouhere.session.exception.SessionIdNotFoundException;
@@ -47,5 +49,10 @@ public class SessionExceptionAdvice {
     @ExceptionHandler(AttendeeNotFoundException.class)
     public ResponseEntity<HttpStatus> attendeeNotFoundHandler(){
         return RESPONSE_BAD_REQUEST;
+    }
+
+    @ExceptionHandler(ActivatedSessionExistsException.class)
+    public ResponseEntity<HttpStatus> activatedSessionExistsHandler(){
+        return RESPONSE_CONFLICT;
     }
 }
